move status bar stuff
